fix(mga): audit track D — postgres:16-alpine + extract source_schemas.py
- docker-compose.yml: postgres:16 → postgres:16-alpine to match MBK canonical
  (same libc/locale/patch cadence between prod DBs)
- Extract SourceCreate, SourceRead, SyncJobResponse out of lineup_schemas.py
  into new app/schemas/game/source_schemas.py (one schema per file convention;
  Source-domain schemas were misfiled in the Lineup schema module)
- Update import sites: api/sources.py, services/game/source_service.py

290 tests pass; 7 classifier setup errors are pre-existing environmental
UniqueViolation (valorant duplicate-game-row, fixed in a separate in-flight PR).
