+++ apps/mygamingassistant/backend/app/schemas/game/lineup_schemas.py
@@ -320,34 +320,3 @@ class PendingLineupsResponse(BaseModel):
     offset: int
 
 
-# ---------------------------------------------------------------------------
-# Source schemas
-# ---------------------------------------------------------------------------
-
-class SourceCreate(BaseModel):
-    kind: str
-    url: str
-
-    @field_validator("kind")
-    @classmethod
-    def validate_kind(cls, v: str) -> str:
-        if v not in ("youtube_playlist", "youtube_channel"):
-            raise ValueError("kind must be youtube_playlist or youtube_channel")
-        return v
-
-
-class SourceRead(BaseModel):
-    id: uuid.UUID
-    kind: str
-    config_json: dict
-    last_synced_at: Optional[str] = None
-    created_at: str
-
-    model_config = {"from_attributes": True}
-
-
-class SyncJobResponse(BaseModel):
-    job_id: str
-    source_id: uuid.UUID
-    status: str = "queued"
-    message: str = "Sync started — lineups will appear in pending_review when complete"
